requirements.txt
Django>=4.2,<5.0
django-parler>=2.3
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
Pillow>=10.0
gunicorn>=21.2
openpyxl>=3.1
